feat(auth-ui): Split pages to separate widgets.
